refactor(uikit): rename Address/utils.ts to Address/AddressType.ts
